reorganize, use s4 for standardized class
trying to do everything through predvars ends up causing some issues
with lsmeans and mixed.  For lme4, it works, but with some warnings.
Instead, now an S4 object of class standardized is created which has a
formula with renamed variables and a corresponding dataframe that can be
used in regression functions.  Also added summary, print, and predict
methods for this class so new data can be placed in the standardized
space before calling other predict methods.  Still need to work on
lsmeans with poly() and scale_by() to ensure that they work properly.