Fixed predict bug; updated documentation
When predicting new standardized data with a mixed effects regression
fit with lme4, there does not appear to be a workaround to prevent
unnecessary warnings.  If the custom contrasts (named_contr_sum and
scaled_contr_poly) are not attached to the factors as attributes in the
newdata frame, there is an error in mkNewReTrms.  If the contrasts are
attached to the factors as attributes, the use of the 'xlev' argument in
predict.merMod when creating the fixed effects model frame causes a
warning that the contrasts are being dropped, but then these same
contrasts are obtained from the 'contrasts' attribute of the model
matrix and reapplied to the factors in the fixed effects frame.  The
result is that warnings are issued about contrasts are being dropped,
but the predictions are correct.